requirements.txt
openai
arize-phoenix
openinference-instrumentation-openai
opentelemetry-api
opentelemetry-sdk
numpy
python-dotenv
tqdm